+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
+* Add custom component for a side panel
 
 ## 2.3.1
 * Fixed destroy -call on an uninitialized Tether component